+++ Code/CrowdControlMod.cs
@@ -951,6 +951,7 @@ public sealed class CrowdControlMod : Mod
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Potion));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Food));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Kite));
+        AddEffect(new SpawnPiggyBankEffect());
 
         // --- World effects
         AddEffect(new UseDialEffect(true));