+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added "Give Yoyo" effect (disabled).
 - Added "Give Accessory" effect (disabled).
+- Added "Spawn Flying Piggy Bank" effect (disabled).
 
 ## [2.0.11] - 2023-12-13